requirements.txt
langchain
langchain-google-genai==2.1.12
google-genai==0.6.0
streamlit==1.39.0
python-dotenv
chromadb
sentence-transformers
pdfplumber